batcher: short-circuit when timeout is 0
There's no point in launching a goroutine and setting a timer and all that junk,
just run the work function inline.

Also clarify that the work function must be concurrency-safe, since that is now
much more obviously true. It was always true, but you could be forgiven for
concluding the opposite on a quick skim of the old code.