Simplify timer cleanup
The timer channels are created with a buffer. Because we aren't reusing them,
we don't need to worry about draining that channel; the internal
goroutine will still be able to write to the buffer and exit gracefully,
at which point the GC will clean it all up.

We *do* still need to call `Stop()` to tell the timer to clean up
immediately, as otherwise it will run for its full duration which (in a
tight loop) can cause timer goroutines to pile up.