Add go 1.22 to CI
It's the latest version, several years newer than the version we were
previously testing, and I expect it to be the one shipped with the next
Ubuntu LTS so it's a good one to test. Keep 1.18 around as it's the
version shipped with the *previous* Ubuntu LTS and so is worth testing
as long as it's supported.